Protect car create/save/delete with admin auth

+++ src/routes/web.php
@@ -36,13 +36,13 @@ Route::delete('/reviews/delete/{id}', 'App\Http\Controllers\ReviewController@del
 // Cars
 Route::get('/cars', 'App\Http\Controllers\CarController@index')->name('car.index');
 
-Route::get('/cars/create', 'App\Http\Controllers\CarController@create')->name('car.create');
+Route::get('/cars/create', 'App\Http\Controllers\CarController@create')->name('car.create')->middleware('auth', 'isAdmin');
 
 Route::get('/cars/{id}', 'App\Http\Controllers\CarController@show')->name('car.show');
 
-Route::post('/cars/save', 'App\Http\Controllers\CarController@save')->name('car.save');
+Route::post('/cars/save', 'App\Http\Controllers\CarController@save')->name('car.save')->middleware('auth', 'isAdmin');
 
-Route::delete('/cars/delete/{id}', 'App\Http\Controllers\CarController@delete')->name('car.delete');
+Route::delete('/cars/delete/{id}', 'App\Http\Controllers\CarController@delete')->name('car.delete')->middleware('auth', 'isAdmin');
 
 // Car model
 Route::get('/car-models', 'App\Http\Controllers\CarModelController@index')->name('carModel.index');
